+++ changelog.md
@@ -2,12 +2,13 @@
 
 ## 2.4.0-beta.1
 
-- Fixed conflict handling for debug keys on 1.21.11.
+- Added a button to copy macros in the profile editor.
 - Added a new activation mode for key Release.
 - Added a command to switch profiles.
 - Added an optional list entry name field for macros.
 - Changed default Conflict Strategy from Submit to Assert.
 - Increased command/message field length limit from 512 to 4096.
+- Fixed conflict handling for debug keys on 1.21.11
 
 ## 2.3.17